Add the slicemetadata.json file into the makerbot file archive
CURA-12005
